// frontend/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        InventoryItemResponse: {
            /** Id */
            id: number;
            /** Shop Item Id */
            shop_item_id: number;
            /** Name */
            name: string;
            /** Category */
            category: string;
            /** Asset Url */
            asset_url: string;
            /** Price Coins */
            price_coins: number;
            /** Rarity */
            rarity: string;
            /**
             * Acquired At
             * Format: date-time
             */
            acquired_at: string;
            /** Placement */
            placement: {
                [key: string]: unknown;
            } | null;
        };
// ...
        PurchaseResponse: {
            /** Success */
            success: boolean;
            /** New Balance */
            new_balance: number;
            item: components["schemas"]["ShopItemResponse"];
        };
// ...
        ShopItemResponse: {
            /** Id */
            id: number;
            /** Name */
            name: string;
            /** Category */
            category: string;
            /** Asset Url */
            asset_url: string;
            /** Price Coins */
            price_coins: number;
            /** Rarity */
            rarity: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}